clear localStorage user when token is expired
