+++ client/src/_helpers/index.js
@@ -1,4 +1,6 @@
 import axios from "axios";
+import jwt_decode from "jwt-decode";
+import router from "../router";
 
 function init_axios(token) {
   axios.defaults.headers.common["Authorization"] = `Bearer ${token}`;
@@ -9,13 +11,28 @@ export function validLink(link) {
 }
 
 export async function initApp(store, token) {
-  init_axios(token);
-  await store.dispatch("students/getStudents");
-  await store.dispatch("logs/getLogs");
-  await store.dispatch("options/getOptions");
+  if (isTokenExpired()) {
+    localStorage.clear();
+    router.push("/login");
+  } else {
+    init_axios(token);
+    await store.dispatch("students/getStudents");
+    await store.dispatch("logs/getLogs");
+    await store.dispatch("options/getOptions");
+  }
 }
 
 export function runGoogleFormProcess(id) {
   const token = JSON.parse(localStorage.getItem("user")).token;
   axios.get(`http://127.0.0.1:8087/${id}?token=${token}`);
+}
+
+function isTokenExpired() {
+  const token =
+    JSON.parse(localStorage.getItem("user")) &&
+    JSON.parse(localStorage.getItem("user"))["token"];
+  if (!token || jwt_decode(token).exp < Date.now() / 1000) {
+    return true;
+  }
+  return false;
 }
